fix: update remaining useStep/useStepper imports after extraction
Three consumer files and step/frame.tsx still imported useStep from
the old barrel re-export path. Updated to import from the new
dedicated hook files.
