+++ client/dashboard/src/components/upload-asset/name-deployment-step.tsx
@@ -3,8 +3,8 @@ import { useLatestDeployment } from "@gram/client/react-query";
 import { Button, Input, Stack } from "@speakeasy-api/moonshine";
 import React from "react";
 import { Type } from "../ui/type";
-import { useStep } from "./step";
-import { useStepper } from "./stepper";
+import { useStep } from "./step/use-step";
+import { useStepper } from "./stepper/use-stepper";
 
 export default function NameDeploymentStep() {
   const stepper = useStepper();
